+++ index.js
@@ -1,13 +1,16 @@
 var array = require('stream-array'),
 	request = require('./lib/util/request'),
 	htmlparser = require('./lib/util/htmlparser'),
+	misc = require('./lib/util/misc'),
+	streams = require('./lib/stream/streams'),
 	PagedReadableStream = require('./lib/stream/paged-readable-stream'),
 	WebpageCrawlingStream = require('./lib/stream/webpage-crawling-stream'),
 	map = require('./lib/map'),
 	enumerate = require('./lib/enumerate'),
 	http = require('./lib/http'),
 	ptt = require('./lib/ptt'),
-	couch = require('./lib/couch');
+	couch = require('./lib/couch'),
+	file = require('./lib/vinyl');
 
 module.exports = {
 	map: map,
@@ -16,12 +19,15 @@ module.exports = {
 	http: http,
 	ptt: ptt,
 	couch: couch,
+	file: file,
 	stream: {
+		endsOf: streams.endsOf,
 		PagedReadableStream: PagedReadableStream,
 		WebpageCrawlingStream: WebpageCrawlingStream
 	},
 	util: {
 		request: request,
-		htmlparser: htmlparser
+		htmlparser: htmlparser,
+		wait: misc.wait
 	}
 };